fix: Split extractors into own files and remove expect()s
