+++ src/infrastructure/http/mod.rs
@@ -3,5 +3,5 @@ pub mod extractors;
 mod server;
 
 pub use error::HttpServerError;
-pub use extractors::{RequestOrigin, UserAgent};
+pub use extractors::{request_origin::RequestOrigin, user_agent::UserAgent};
 pub use server::HttpServer;
